Extract helper class
